+++ packages/astro/scripts/build.js
@@ -1,11 +1,11 @@
 import assert from 'node:assert';
-import path from 'node:path';
 import { existsSync, rmSync, copyFileSync } from 'node:fs';
 import { cp, rm } from 'node:fs/promises';
-import { execa } from 'execa';
+import path from 'node:path';
 import chokidar from 'chokidar';
 import esbuild from 'esbuild';
 import { nodeExternalsPlugin } from 'esbuild-node-externals';
+import { execa } from 'execa';
 
 const isWatch = process.argv.includes('--watch');
 
@@ -46,6 +46,7 @@ async function buildJS() {
     outdir: 'dist',
     define: {
       'process.env.TUTORIALKIT_DEV': JSON.stringify(process.env.TUTORIALKIT_DEV ?? null),
+      'process.env.TUTORIALKIT_VITE_INSPECT': JSON.stringify(process.env.TUTORIALKIT_VITE_INSPECT ?? null),
     },
     plugins: [nodeExternalsPlugin()],
   });
@@ -64,14 +65,11 @@ async function copyDefaultFolder() {
   await rm(dist, { recursive: true, force: true });
 
   // copy default folder unmodified, without test files
-  await cp(src, dist, {
-    recursive: true,
-    filter: (filename) => !filename.endsWith('.spec.ts'),
-  });
+  await cp(src, dist, { recursive: true, filter });
 
   if (isWatch) {
     chokidar.watch(src).on('all', (event, filePath, stats) => {
-      if (stats?.isDirectory() !== true) {
+      if (stats?.isDirectory() !== true && filter(filePath)) {
         const target = path.join(dist, path.relative(src, filePath));
 
         if (event === 'unlink') {
@@ -81,5 +79,9 @@ async function copyDefaultFolder() {
         }
       }
     });
+  }
+
+  function filter(filename) {
+    return !filename.endsWith('.spec.ts') && !filename.includes('__snapshots__');
   }
 }