+++ packages/types/src/entities/index.ts
@@ -1,9 +1,13 @@
 import type { I18nSchema } from '../schemas/i18n.js';
 import type { ChapterSchema, LessonSchema, PartSchema } from '../schemas/index.js';
+import type { MetaTagsSchema } from '../schemas/metatags.js';
 
 export type * from './nav.js';
 
+export type FileDescriptor = { path: string; type: 'file' | 'folder' };
 export type Files = Record<string, string | Uint8Array>;
+
+export type FilesystemError = 'FILE_EXISTS' | 'FOLDER_EXISTS';
 
 /**
  * This tuple contains a "ref" which points to a file to fetch with the `LessonFilesFetcher` and
@@ -53,6 +57,8 @@ export interface Lesson<T = unknown> {
 }
 
 export type I18n = Required<NonNullable<I18nSchema>>;
+
+export type MetaTagsConfig = MetaTagsSchema;
 
 export interface Tutorial {
   logoLink?: string;