+++ packages/types/src/schemas/common.ts
@@ -1,5 +1,6 @@
 import { z } from 'zod';
 import { i18nSchema } from './i18n.js';
+import { metaTagsSchema } from './metatags.js';
 
 export const commandSchema = z.union([
   // a single string, the command to run
@@ -54,6 +55,15 @@ export const previewSchema = z.union([
 ]);
 
 export type PreviewSchema = z.infer<typeof previewSchema>;
+
+export const fileSystemSchema = z.object({
+  watch: z
+    .boolean()
+    .optional()
+    .describe('When set to true, file changes in WebContainer are updated in the editor as well.'),
+});
+
+export type FileSystemSchema = z.infer<typeof fileSystemSchema>;
 
 const panelTypeSchema = z
   .union([z.literal('output'), z.literal('terminal')])
@@ -161,10 +171,45 @@ export const terminalSchema = z.union([
   }),
 ]);
 
+export const editorSchema = z.union([
+  // can either be completely removed by setting it to `false`
+  z.boolean().optional(),
+
+  z.strictObject({
+    fileTree: z
+      .union([
+        // or you can only remove the file tree
+        z.boolean(),
+
+        // or configure file tree with options
+        z.strictObject({
+          allowEdits: z
+            .union([
+              // allow editing all files or disable completely
+              z.boolean(),
+
+              // limit file editing to files and folders that match a single glob pattern
+              z.string(),
+
+              // limit file editing to files and folders that match one of multiple glob patterns
+              z.array(z.string()),
+            ])
+            .describe(
+              'Allow file tree’s items to be edited by right clicking them. Supports file and folder creation.',
+            ),
+        }),
+      ])
+      .optional(),
+  }),
+]);
+
 export type TerminalPanelType = z.infer<typeof panelTypeSchema>;
 export type TerminalSchema = z.infer<typeof terminalSchema>;
+export type EditorSchema = z.infer<typeof editorSchema>;
 
 export const webcontainerSchema = commandsSchema.extend({
+  meta: metaTagsSchema.optional(),
+
   previews: previewSchema
     .optional()
     .describe(
@@ -175,6 +220,11 @@ export const webcontainerSchema = commandsSchema.extend({
     .optional()
     .describe(
       'Navigating to a lesson that specifies autoReload will always reload the preview. This is typically only needed if your server does not support HMR.',
+    ),
+  filesystem: fileSystemSchema
+    .optional()
+    .describe(
+      'Configure how changes happening on the filesystem should impact the Tutorial. For instance, when new files are being changed, whether those change should be reflected in the editor.',
     ),
   template: z
     .string()
@@ -191,18 +241,10 @@ export const webcontainerSchema = commandsSchema.extend({
     .string()
     .optional()
     .describe('Defines which file should be opened in the code editor by default when lesson loads.'),
-  editor: z
-    .union([
-      // can either be completely removed by setting it to `false`
-      z.boolean().optional(),
-
-      // or you can only remove the file tree
-      z.strictObject({
-        fileTree: z.boolean().optional(),
-      }),
-    ])
+  editor: editorSchema
+    .optional()
     .describe(
-      'Configure whether or not the editor should be rendered. If an object is provided with fileTree: false, only the file tree is hidden.',
+      'Configure whether or not the editor should be rendered. File tree can be configured by proving an object with fileTree option.',
     ),
   i18n: i18nSchema
     .optional()